+++ app/src/main/java/sg/edu/nus/iss/se/ft05/medipal/fragments/IceFragment.java
@@ -71,7 +71,7 @@ public class IceFragment extends Fragment {
                 // COMPLETED (8) Inside, get the viewHolder's itemView's tag and store in a long variable id
                 //get the id of the item being swiped
 
-                long id = (long) viewHolder.itemView.getTag();
+                int id = (int) viewHolder.itemView.getTag();
 
                 // COMPLETED (9) call removeGuest and pass through that id
                 //remove from DB
